+++ client/src/components/classes/Dashboard.tsx
@@ -1,15 +1,15 @@
 import moment from "moment";
-import React, { useContext, useEffect, useState } from "react";
-import linkimg from '../../assets/link.png';
+import { useContext, useEffect, useState } from "react";
 import CurrentTimeContext from "../../contexts/CurrentTimeContext";
 import UserDataContext, { SgyData, SgyPeriodData, UserData } from "../../contexts/UserDataContext";
 
-import { parsePeriodColor, SCHOOL_END_EXCLUSIVE, SCHOOL_START } from "../schedule/Periods";
-import { Functions, httpsCallable } from 'firebase/functions';
-import { useAuth, useFunctions } from "reactfire";
-import { fetchSgyMaterials, findClassesList } from "../../views/Classes";
-import { getSchedule, useSchedule } from "../../hooks/useSchedule";
+import { parsePeriodColor } from "../schedule/Periods";
+import { findClassesList } from "../../views/Classes";
 import { useScreenType } from "../../hooks/useScreenType";
+import { UpcomingQuickCal } from "./dashboard/QuickWeekCal";
+import { getAllGrades, getUpcomingInfo } from "./functions/SgyFunctions";
+import { ClassPeriodQuickInfo, pastClasses, nextSchoolDay, numSchoolDays } from "./functions/PeriodFunctions";
+import { cardinalize, classifyGrade } from "./functions/GeneralHelperFunctions";
 
 export type DashboardAssignment = {
     name: string;
@@ -17,50 +17,6 @@ export type DashboardAssignment = {
     timestamp: moment.Moment;
     description:string;
     period: string;
-}
-
-const UpcomingQuickCalDot = (props: {course:string}) => {
-    const userData = useContext(UserDataContext);
-    return <div className="upcoming-blurb-qc-dot" style={{backgroundColor:parsePeriodColor(props.course, userData)}}></div>
-}
-
-const UpcomingQuickCalDay = (props: { day: moment.Moment, upcoming: DashboardAssignment[], selected:string}) => {
-
-    const { day, upcoming, selected } = props;
-    const screenType = useScreenType();
-
-    const weekdays = ['U', 'M', 'T', 'W', 'θ', 'F', 'S']
-
-    const relevantAssigments = upcoming.filter((a) => a.timestamp.dayOfYear() === day.dayOfYear());
-
-    let active = false;
-    if(selected === 'A') {
-        if(getSchedule(day)) active = true;
-    } else {
-        if(hasClass(day, selected)) active = true;
-    }
-
-    return <div className={"upcoming-blurb-qc-day" + (active ? '-active' : '-inactive' )}>
-        <div className="upcoming-blurb-qc-day-num"> {screenType === 'phone' ? weekdays[day.weekday()] : `${weekdays[day.weekday()]} • ${day.date()}`}</div>
-        <div className="upcoming-blurb-qc-dots">
-            {relevantAssigments.map((a) => <UpcomingQuickCalDot key={a.period} course={a.period} />)}
-        </div>
-    </div>
-}
-
-const UpcomingQuickCal = (props: { upcoming: DashboardAssignment[], selected:string }) => {
-    const {upcoming,selected} = props;
-    const time = useContext(CurrentTimeContext);
-    let mutableTime = moment(time);
-    const days = [];
-    for(let i = 0; i < 7; i++) {
-        days.push(moment(mutableTime));
-        mutableTime.add(1, "days");
-    }
-
-    return <div className="upcoming-blurb-quick-cal">
-        {days.map((day) => <UpcomingQuickCalDay key={day.format('YYYY-MM-DD')} selected={selected} day={day} upcoming={upcoming} />)}
-    </div>
 }
 
 const UBAssignment = (props:{name:string, due:string, period:string}) => {
@@ -76,13 +32,9 @@ const UBAssignment = (props:{name:string, due:string, period:string}) => {
 }
 
 const UBAssignments = (props: { upcoming: DashboardAssignment[] }) => {
-
     const upcoming = props.upcoming.slice(0,5); // only display up to 5
 
     return <div>
-        {/* <UBAssignment name={"Activity Series of Metals Lab - Get your document here"} due={"Wednesday, December 1st • In 5 days"} />
-        {Array(2).fill(<UBAssignment name={"e"} due={"Wednesday, December 1st • In 5 days"} />)} */}
-
         {upcoming.map((a) => <UBAssignment key={a.link} name={a.name} due={`${a.timestamp.format("dddd, MMMM Do")} • ${a.timestamp.fromNow()}`} period={a.period} />)}
         <div className="ub-upcoming-redirect"><div>See More in Upcoming</div></div>
     </div>
@@ -114,308 +66,9 @@ const DashLeftSection = (props: { upcoming: DashboardAssignment[] | null, select
     </div>
 }
 
-const hasClass = (day:moment.Moment, period:string) => {
-    return getSchedule(day)?.find(([p]) => p === period) ?? null;
-}
-
-const findNextClass = (period: string) => {
-    const now = moment();
-
-    while (!hasClass(now, period) && !now.isAfter(SCHOOL_END_EXCLUSIVE)) { // while the schedule doesn't have the period and we're still in the school year
-        now.add(1,'days'); // increment day
-    }
-
-    if (now.isAfter(SCHOOL_END_EXCLUSIVE) ) return null;
-
-    const p = hasClass(now, period)!;
-    if(p) {
-        const t = p[1].s;
-        const m = t % 60;
-        const h = (t-m) / 60;
-
-        now.set("hour", h);
-        now.set("minute", m);
-        now.set("second", 0);
-
-        return now;
-    }
-
-    return null;
-}
-
-type ClassQuickInfo = {
-    next?: {
-        time: moment.Moment;
-        week: number;
-        day: number;
-    }
-    past: {
-        days: number;
-    }
-}
-
-
-const pastClasses = (period: string): ClassQuickInfo => {
-    const current = moment(SCHOOL_START);
-    while (!hasClass(current, period) && !current.isAfter(SCHOOL_END_EXCLUSIVE)) current.add(1, 'days'); // find first instance of class
-
-    let weeks = 1;
-    let day = 1;
-    let days = 1;
-
-    let prev = moment(current);
-
-    while (current.isBefore(moment()) && !current.isAfter(SCHOOL_END_EXCLUSIVE)) {
-        current.add(1, 'days');
-
-        if (hasClass(current, period)) {
-            days++;
-
-            if (!current.isSame(prev, 'week')) {// if it's not in the same week
-                // new week!
-                weeks++;
-                day = 1;
-                prev = moment(current);
-            } else {
-                day++;
-            }
-        }
-    }
-
-    let next = findNextClass(period);
-
-    if(next) {
-        if(!next.isSame(prev,'week')) {
-            weeks++;
-            day = 1;
-        } else {
-            day++;
-        }
-
-        return {
-            next: {
-                time: next!,
-                week: weeks,
-                day
-            },
-            past: {
-                days
-            }
-        }
-    }
-
-    return {
-        past: {
-            days
-        }
-    }
-}
-
-const nextSchoolDay = () => {
-    const now = moment();
-
-    while (!getSchedule(now) && !now.isAfter(SCHOOL_END_EXCLUSIVE)) { 
-        now.add(1, 'days'); // increment day
-    }
-
-    if (now.isAfter(SCHOOL_END_EXCLUSIVE)) return null;
-
-    const p = getSchedule(now)![0];
-    if (p) {
-        const t = p[1].s;
-        const m = t % 60;
-        const h = (t - m) / 60;
-
-        now.set("hour", h);
-        now.set("minute", m);
-        now.set("second", 0);
-
-        return now;
-    }
-
-    return null;
-}
-
-const numSchoolDays = () => {
-    const current = moment(SCHOOL_START);
-
-    let days = 1;
-
-    while (current.isBefore(moment()) && !current.isAfter(SCHOOL_END_EXCLUSIVE)) {
-        current.add(1, 'days');
-
-        if(getSchedule(current)) {
-            days++;
-        }
-    }
-    return days;
-}
-
-const cardinalize = (num:number) => {
-    switch(num%10) {
-        case 1:
-            return num + 'st';
-        case 2:
-            return num + 'nd';
-        case 3:
-            return num + 'rd';
-        default:
-            return num + 'th';
-    }
-}
-
-const momentComparator = (a: moment.Moment, b: moment.Moment) => {
-    if (a.isBefore(b)) {
-        return -1;
-    }
-    if (a.isAfter(b)) {
-        return 1;
-    }
-    return 0;
-}
-
-const findGrades = (sgyData: SgyData, selected: string) => {
-    const selectedCourse = sgyData[selected];
-    // Attempt to match the id of the selected course to the id in the course grades
-    let selectedCourseGrades = sgyData.grades.find(sec => sec.section_id === selectedCourse.info.id);
-
-    if (!selectedCourseGrades) {
-        // they do this quirky and uwu thing where THEY CHANGE THE ID
-
-        // the way we match this is by searching through the courses and seeing if the assignments match up, which is so stupid that it works
-        for (const course in sgyData.grades) {
-            if (sgyData.grades[course].period[0].assignment.length > 0) {
-                const assiToFind = sgyData.grades[course].period[0].assignment.find(assignment => assignment.type === 'assignment');
-                if (assiToFind && selectedCourse.assignments.find(assi => assi.id === assiToFind.assignment_id)) {
-                    // BANANA
-                    selectedCourseGrades = sgyData.grades[course];
-                    break;
-                }
-            }
-        }
-    }
-
-    return selectedCourseGrades || null;
-
-}
-
-const getAllGrades = (sgyData: SgyData, userData: UserData) => {
-    const classes = findClassesList(userData);
-    const grades: { [key: string]: number } = {};
-
-    for (const c of classes) {
-        if (c.period === "A") continue;
-
-        const selectedCourseGrades = findGrades(sgyData, c.period); // find the grades
-        if (selectedCourseGrades) grades[c.period] = selectedCourseGrades.final_grade[0].grade;
-    }
-
-    return grades;
-}
-
-export const getUpcomingInfo = (sgyData: SgyData, selected: string, userData: UserData, time: moment.Moment) => {
-
-    if (selected === 'A') {
-        const upcoming: DashboardAssignment[] = [];
-        const overdue: DashboardAssignment[] = [];
-        // const grades: { [key: string]: number } = {};
-
-        const classes = findClassesList(userData);
-
-        for (const c of classes) {
-            if (c.period === "A") continue;
-            const courseStuff = getUpcomingInfo(sgyData, c.period, userData, time);
-            if (courseStuff) {
-                upcoming.push(...courseStuff.upcoming);
-                overdue.push(...courseStuff.overdue);
-                // if (courseStuff.finalGrade) grades[c.period] = courseStuff.finalGrade;
-            }
-        }
-
-        upcoming.sort((a, b) => momentComparator(a.timestamp, b.timestamp));
-        overdue.sort((a, b) => momentComparator(a.timestamp, b.timestamp));
-
-        return { upcoming, overdue };
-    }
-
-    // Select the course
-    const selectedCourse = sgyData[selected];
-    const selectedCourseGrades = findGrades(sgyData, selected); // find the grades
-
-    const upcoming: DashboardAssignment[] = [];
-    const overdue: DashboardAssignment[] = [];
-
-    // search thru assignments
-    for (const item of selectedCourse.assignments) {
-        // console.log(item.title, item.grade_stats);
-        if (item.due.length > 0) { // if it's actually due
-
-            const due = moment(item.due);
-
-            if (due.isAfter(time)) { // if it's due after right now
-
-                // format the assignment
-                const assi: DashboardAssignment = {
-                    name: item.title,
-                    description: item.description,
-                    link: `https://pausd.schoology.com/assignment/${item.id}`,
-                    timestamp: moment(item.due),
-                    period: selected
-                }
-                upcoming.push(assi);
-            } else {
-                // check if it's overddue
-                if (selectedCourseGrades) {
-                    let found = false;
-                    for (const period of selectedCourseGrades.period) {
-                        for (const assi of period.assignment) {
-                            if (assi.assignment_id === item.id) {
-                                found = true;
-                            }
-                        }
-                    }
-
-                    if (!found) {
-                        overdue.push({
-                            name: item.title,
-                            description: item.description,
-                            link: `https://pausd.schoology.com/assignment/${item.id}`,
-                            timestamp: moment(item.due),
-                            period: selected
-                        });
-                    }
-                }
-            }
-        }
-    }
-
-    // do the same for events
-    for (const item of selectedCourse.events) {
-        const due = moment(item.start);
-
-        if (due.isAfter(time) && !item.assignment_id) {
-            const assi: DashboardAssignment = {
-                name: item.title,
-                description: item.description,
-                link: `https://pausd.schoology.com/event/${item.id}`,
-                timestamp: moment(item.start),
-                period: selected
-            }
-            upcoming.push(assi);
-        }
-    }
-
-    upcoming.sort((a, b) => momentComparator(a.timestamp, b.timestamp));
-    overdue.sort((a, b) => momentComparator(a.timestamp, b.timestamp));
-
-    const finalGrade = selectedCourseGrades ? selectedCourseGrades.final_grade[0].grade : null;
-
-    return { upcoming, overdue };
-}
-
 const DashboardQuickInfo = (props:{selected:string}) => {
     const {selected} = props;
-    const [info, setInfo] = useState<ClassQuickInfo | null>(null);
+    const [info, setInfo] = useState<ClassPeriodQuickInfo | null>(null);
 
     useEffect(() => {
         if(selected !== 'A') setInfo(pastClasses(selected));
@@ -442,20 +95,6 @@ const DashboardQuickInfo = (props:{selected:string}) => {
         <div className={"dashboard-qi-main"}>The next class is {info.next?.time.fromNow()}.</div>
         <div className={"dashboard-qi-note"}>It will be on {info.next?.time.format('dddd, MMMM Do')}, and will be Week {info.next?.week} Day {info.next?.day}, the {cardinalize(info.past.days+1)} class of the school year. </div>
     </>
-}
-
-const classifyGrade = (grade:number) => {
-    if(grade >= 93) return 'A';
-    if(grade >= 90) return 'A-';
-    if(grade >= 87) return 'B+';
-    if(grade >= 83) return 'B';
-    if(grade >= 80) return 'B-';
-    if(grade >= 77) return 'C+';
-    if(grade >= 73) return 'C';
-    if(grade >= 70) return 'C-';
-    if(grade >= 67) return 'D+';
-    if(grade >= 63) return 'D';
-    return 'D-';
 }
 
 const DashGrades = (props: { selected:string, allGrades: {[key:string]:number} }) => {
