+++ client/src/components/classes/functions/GeneralHelperFunctions.ts
@@ -1,0 +1,27 @@
+
+export const classifyGrade = (grade: number) => {
+    if (grade >= 93) return 'A';
+    if (grade >= 90) return 'A-';
+    if (grade >= 87) return 'B+';
+    if (grade >= 83) return 'B';
+    if (grade >= 80) return 'B-';
+    if (grade >= 77) return 'C+';
+    if (grade >= 73) return 'C';
+    if (grade >= 70) return 'C-';
+    if (grade >= 67) return 'D+';
+    if (grade >= 63) return 'D';
+    return 'D-';
+}
+
+export const cardinalize = (num: number) => {
+    switch (num % 10) {
+        case 1:
+            return num + 'st';
+        case 2:
+            return num + 'nd';
+        case 3:
+            return num + 'rd';
+        default:
+            return num + 'th';
+    }
+}
